+++ src/main/java/com/linerup/lineup_backend/security/component/JwtAuthenticationFilter.java
@@ -1,0 +1,29 @@
+package com.linerup.lineup_backend.security.component;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+/**
+ * @author : hyunwoopark
+ * @version : 1.0.0
+ * @package : LineUP_BackEnd
+ * @name : JwtAuthenticationFilter
+ * @date : 2023/09/01 5:38 PM
+ * @modifyed : $
+ **/
+@Component
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+    }
+}
