+++ src/main/java/com/linerup/lineup_backend/security/oauth2/userinfo/CustomOAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.linerup.lineup_backend.oauth2.userinfo;
+package com.linerup.lineup_backend.security.oauth2.userinfo;
 /**
 * @package : com.linerup.lineup_backend.oauth2.userinfo
 * @name : CustomOAuth2UserInfo.java
