+++ src/main/java/com/linerup/lineup_backend/security/oauth2/userinfo/GithubOAuth2UserInfo.java
@@ -1,25 +1,25 @@
-package com.linerup.lineup_backend.oauth2.userinfo;
+package com.linerup.lineup_backend.security.oauth2.userinfo;
 /**
 * @author : 
 * @version : 1.0.0
 * @package : com.linerup.lineup_backend.oauth2.userinfo
-* @name : GoogleOAuth2UserInfo.java
-* @date : 2023/08/25 4:44 AM
+* @name : GithubOAuth2UserInfo.java
+* @date : 2023/08/25 4:41 AM
 * @modified : 
 **/
 import java.util.Map;
 
-public class GoogleOAuth2UserInfo implements CustomOAuth2UserInfo {
+public class GithubOAuth2UserInfo implements CustomOAuth2UserInfo {
 
   private final Map<String, Object> attributes;
 
-  public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+  public GithubOAuth2UserInfo(Map<String, Object> attributes) {
     this.attributes = attributes;
   }
 
   @Override
   public String getProviderId() {
-    return attributes.get("sub").toString();
+    return String.valueOf(attributes.get("id"));
   }
 
   @Override
